harness/requirements.txt
pyyaml==6.0.1
jsonschema==4.21.1
pytest==8.1.1
rich==13.7.1

